Move wireless files, removed wireless command
