+++ src/lib/connect/windows.js
